Adding file with tests for raincloud plots
When writing functions it is good practice to include a test file. Formal tests can be done with the testthat package but for plots it doesn't help much as you want to see what they look like.

It is good to include a test that runs all of the sub functions, starting with the most basic so you can see how they all run.