Refactoring check_values to get rid of nested loop + test script
I created functions for each level of the loop and used pmap and apply to implement the loop. It took me a bit of time to do and I'm not certain tons clearer than the nested loop :).

In the test script I use testthat. See https://r-pkgs.org/testing-basics.html.
  It is more intended for use when developing packages so is probably overkill here.