Adding optional argument of variable metadata to make raincloud plots function
If the variable meta data is suplied the make rainclouds plot function will use this to find the min max values to use in the scales.

This funciton makes use of pmap and a tibble of arguments - I find this to be a great way to loop over many things at once.